backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sse-starlette==2.1.3
httpx==0.27.0
transformers==4.44.0
torch==2.3.0
sentencepiece==0.2.0
pydantic==2.8.0
python-dotenv==1.0.1
protobuf==5.28.0
